v2.4.1.1 -> v2.4.1.2
  - driver sync up with Alan
  - Andrew Morton: wakeup cleanup and race fix
  - Paul Mackerras: macintosh driver updates.
  - don't trust "page_count()" on reserved pages!
  - Russell King: fix serious IDE multimode write bug!
  - me, Jens, others: fix elevator problem
  - ARM, MIPS and cris architecture updates
  - alpha updates: better page clear/copy, avoid kernel lock in execve
  - USB and firewire updates
  - ISDN updates
  - Irda updates

+++ drivers/acorn/block/fd1772.c
@@ -276,8 +276,8 @@ static int MotorOn = 0, MotorOffTrys;
 static volatile int fdc_busy = 0;
 static DECLARE_WAIT_QUEUE_HEAD(fdc_wait);
 
-
-static unsigned int changed_floppies = 0xff, fake_change = 0;
+/* long req'd for set_bit --RR */
+static unsigned long changed_floppies = 0xff, fake_change = 0;
 #define	CHECK_CHANGE_DELAY	HZ/2
 
 /* DAG - increased to 30*HZ - not sure if this is the correct thing to do */
